Breaking up ABRIK into two components
