+++ benchmark/bench_ABRIK/ABRIK_runtime_breakdown.cc
@@ -65,8 +65,6 @@ static void call_all_algs(
     // Additional params setup.
     RandLAPACK::ABRIK<T, r123::Philox4x32> ABRIK(false, time_subroutines, tol);
     ABRIK.max_krylov_iters = num_matmuls;
-    ABRIK.num_threads_min = 4;
-    ABRIK.num_threads_max = RandLAPACK::util::get_omp_threads();
 
     auto state_alg = state;
     std::vector<long> inner_timing;
